+++ firmware/esp32s3-usb-kbd/main/keymap.h
@@ -6,4 +6,4 @@
 // key_id (0..255) -> (modifier, keycode)
 // Left controller uses 1..35, right controller uses 129..163 (128 + base).
 // Returns false if unmapped.
-bool keymap_lookup(uint8_t key_id, uint8_t* modifier, uint8_t* keycode);
+bool keymap_lookup(uint16_t key_id, uint8_t* modifier, uint8_t* keycode);
